Linear Loss Function

Compute an approximation of the integral of the function f(x) with respect to dx in the range [a, b] by Monte-Carlo integration using uniform sampling.

beta_ls(resp, pred, beta, norm = "L2")

Arguments

resp

A vector of dimension 2 used to denote the integration region of interest, i.e. [a, b].

pred

A string containing the function to be integrated. It is assumed that x is used as the variable of interest.

beta

A numeric (integer) used to denote the number of simulations.

norm

A numeric used to control the seed of the random number generator used by this function.

Value

A numeric containing the loss function at beta

Author

Rob Molinari

Examples

set.seed(123)
X <- matrix(rnorm(600), nrow = 200) # true matrix of predictors
epsilon <- rnorm(200, 0, sd = 0.25) # observation error
X <- cbind(rep(1, 200), X) # add a column of ones to generate the response with an intercept value
beta <- c(-1, 4, -5, 2) # true coefficient values we want to estimate (first element is the intercept)
y <- X%*%beta + epsilon # generate response y based on this model
beta_ls(resp = y, pred = X, beta = beta, norm = "L2")
#>          [,1]
#> [1,] 13.06387